结构改变

+++ database/migrations/2018_04_25_104724_Rbac.php
@@ -30,7 +30,7 @@ class Rbac extends Migration
             $table->string( 'description' )->nullable();
             $table->tinyInteger( 'is_menu' )->default( 0 )->comment( '是否作为菜单显示,[1|0]' );
             $table->tinyInteger( 'init_curd' )->default( 0 )->comment( '是否初始化增删改权限节点,[1|0]' );
-            $table->tinyInteger( 'sort' )->default( 0 )->comment( '排序' );
+            $table->integer( 'sort' )->default( 0 )->comment( '排序' );
             $table->timestamps();
         } );
